Trying to fix product list type color

+++ be/src/main/java/vn/fpt/diamond_shop/service/IJewelryService.java
@@ -11,7 +11,7 @@ import java.util.Optional;
 public interface IJewelryService {
     Jewelry getJewelryById(Long id);
     List<Jewelry> getAllJewelries();
-    List<Jewelry> getJewelriesByFilter(List<EJewelryType> types, Double minPrice, Double maxPrice);
+    List<Jewelry> getJewelriesByFilter(List<EJewelryType> types, Integer minPrice, Integer maxPrice);
     Integer calculateJewelryPrice(Jewelry jewelry);
     Integer calculateJewelryPriceWithSize(Jewelry jewelry, JewelrySize size);
     void deleteJewelryById(Long id);
